+++ app/test/e2e/v1/get-fw.spec.js
@@ -75,6 +75,7 @@ describe('V1 - Get FW areas tests', () => {
         response.body.data[0].attributes.should.have.property('wdpaid').and.equal(area.wdpaid);
         response.body.data[0].attributes.should.have.property('userId').and.equal(USERS.USER.id);
         response.body.data[0].attributes.should.have.property('createdAt');
+        response.body.data[0].attributes.should.have.property('updatedAt');
         response.body.data[0].attributes.should.have.property('image').and.equal('');
         response.body.data[0].attributes.should.have.property('datasets').and.be.an('array').and.length(0);
         response.body.data[0].attributes.should.have.property('use').and.be.an('object');
@@ -138,6 +139,7 @@ describe('V1 - Get FW areas tests', () => {
         response.body.data[0].attributes.should.have.property('wdpaid').and.equal(area.wdpaid);
         response.body.data[0].attributes.should.have.property('userId').and.equal(area.userId);
         response.body.data[0].attributes.should.have.property('createdAt');
+        response.body.data[0].attributes.should.have.property('updatedAt');
         response.body.data[0].attributes.should.have.property('image').and.equal('');
         response.body.data[0].attributes.should.have.property('datasets').and.be.an('array').and.length(0);
         response.body.data[0].attributes.should.have.property('use').and.be.an('object');
@@ -257,6 +259,7 @@ describe('V1 - Get FW areas tests', () => {
         response.body.data[0].attributes.should.have.property('wdpaid').and.equal(area.wdpaid);
         response.body.data[0].attributes.should.have.property('userId').and.equal(area.userId);
         response.body.data[0].attributes.should.have.property('createdAt');
+        response.body.data[0].attributes.should.have.property('updatedAt');
         response.body.data[0].attributes.should.have.property('image').and.equal('');
         response.body.data[0].attributes.should.have.property('datasets').and.be.an('array').and.length(0);
         response.body.data[0].attributes.should.have.property('use').and.be.an('object');
@@ -303,6 +306,7 @@ describe('V1 - Get FW areas tests', () => {
         response.body.data[0].attributes.should.have.property('wdpaid').and.equal(area.wdpaid);
         response.body.data[0].attributes.should.have.property('userId').and.equal(area.userId);
         response.body.data[0].attributes.should.have.property('createdAt');
+        response.body.data[0].attributes.should.have.property('updatedAt');
         response.body.data[0].attributes.should.have.property('image').and.equal('');
         response.body.data[0].attributes.should.have.property('datasets').and.be.an('array').and.length(0);
         response.body.data[0].attributes.should.have.property('use').and.be.an('object');